feat: Move AppSettingsViewModel to shared usecase module (Phase 28)
Extract Setting<T> and AppSettingsRepository interfaces to domain.
Move DefaultAppSettingsViewModel to usecase/src/commonMain/.
Android AppSettings extends domain AppSettingsRepository.
