+++ AndroidGarage/androidApp/src/main/java/com/chriscartland/garage/settings/AppSettings.kt
@@ -24,6 +24,7 @@ import com.chriscartland.garage.settings.SettingType.LongSetting
 import com.chriscartland.garage.settings.SettingType.StringSetting
 
 interface AppSettings :
+    com.chriscartland.garage.domain.repository.AppSettingsRepository,
     SettingContract,
     SettingManager
